Make more of Krop cross platform
I realised that a lot more of http4s is cross-platform that I expected,
so make more of Krop cross-platform as a result.
