+++ core/jvm/src/main/scala/krop/syntax/all.scala
@@ -16,4 +16,4 @@
 
 package krop.syntax
 
-object all extends RouteSyntax {}
+object all {}
