+++ core/shared/src/main/scala/krop/Logger.scala
