+++ core/shared/src/main/scala/krop/Mode.scala
