+++ core/shared/src/main/scala/krop/tool/KropAssets.scala
